remove initial memset of values in parquet reader (#19643)
In the parquet reader we pre-fill the output column data/offset buffers with zeroes to make sure we don't have any uninitialized reads.  This is slow and seems to be unnecessary, because the parquet reader should write over that memory anyway when reading the data.  However, the reader doesn't write to the value buffer's memory where there are nulls, and without zero-initialization, reading this memory leads to crashes.  

When do we read nulled values/offsets?  You'd think it'd be never, but we do so when we: 
* check_non_empty_nulls() 
* Trying to debug-print a column to screen
* Trying to compare/print any differences between expected/result in tests
* element<string_view>() reads AND USES invalid offsets
* etc. 

So we still remove the pre-zeroing, but instead write zeroes in data_out in the parquet reader decode functions for the entries that are null.  This makes it so we only spend time zeroing what we need to (which is probably sparse) rather than the entire column. 

Also, because we write the input int32 TIME_MILLIS to an output int64 DURATION_MILLISECONDS, we have to fill in the last 4 bytes of zeroes. 

The performance improvement on my machine is about 5% for most column types, though is negligible for strings (because per is dominated by the string copy, not the offset write).  Further work can be done to try to remove the null mask zeroing as well.

Authors:
  - Paul Mattione (https://github.com/pmattione-nvidia)

Approvers:
  - https://github.com/nvdbaranec
  - Vukasin Milovanovic (https://github.com/vuule)

URL: https://github.com/rapidsai/cudf/pull/19643