Update exception handling in pdsh benchmarks (#19793)
This updates how we handle exceptions in the execution of a cudf-polars query. Previously, an exception raised during a query (e.g. an out of memory exception) would cause the process to exit, and no results would be written.

Now, we'll

1. Catch the exception, print it to stdout, and record the fact that that query-iteration errored
2. Continue running the remaining queries / iterations
3. Write out the results of any sucessful execution
4. Exit with a non-zero error code if any exceptions occurred

Authors:
  - Tom Augspurger (https://github.com/TomAugspurger)

Approvers:
  - Richard (Rick) Zamora (https://github.com/rjzamora)
  - Matthew Murray (https://github.com/Matt711)

URL: https://github.com/rapidsai/cudf/pull/19793