Move prefetching out of experimental and simplify the API (#19875)
We have been using prefetching in cudf.pandas for almost a year now, and for a shorter time in cudf-polars, both without issue. We have not modified the places where we prefetch or how we use the APIs in a long time. This PR therefore moves the libcudf prefetching logic out of an experimental namespace. It also simplifies the APIs for enabling prefetching, removing per-key control and instead making it an all-or-nothing choice.

I've marked the PR as breaking, but since we are changing experimental features we do not need to provide any sort of compatibility shim.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Nghia Truong (https://github.com/ttnghia)
  - David Wendt (https://github.com/davidwendt)
  - Matthew Roeschke (https://github.com/mroeschke)

URL: https://github.com/rapidsai/cudf/pull/19875