Coalesce IO of chunks with different compression when reading Parquet files (#19884)
issue https://github.com/rapidsai/cudf/issues/17313

Parquet reader used to deallocate compressed chunks before decode to reduce memory use. Because of this, IO of compressed and uncompressed chunks is not coalesced together.

Since we introduced chunking in the reader, the compressed chunks are not deallocated early and we can safely coalesce across chunks regardless of their compression.

This PR removes the compression condition when coalescing IO, leading to better IO performance for files with mixed chunks.

If we decide to go back to deallocating the compressed chunks early, we need to restore this condition.

Authors:
  - Vukasin Milovanovic (https://github.com/vuule)

Approvers:
  - Muhammad Haseeb (https://github.com/mhaseeb123)
  - Basit Ayantunde (https://github.com/lamarrr)

URL: https://github.com/rapidsai/cudf/pull/19884